requirements.txt
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.100.0
google-auth==2.26.1
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
tensorflow-cpu==2.14.0
torch==2.1.1
transformers==4.35.2
scikit-learn==1.3.2
textblob==0.17.1
nltk==3.8.1
deap==1.4.1
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
joblib==1.3.2
tqdm==4.66.1
